requirements.txt
# Core LangChain packages
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-vertexai==2.0.28
langchain-text-splitters==0.3.11

# Document processing
pypdf==6.0.0
docx2txt==0.9
python-docx==1.2.0

# Vector database
chromadb==1.0.20

# Google Cloud dependencies
google-cloud-aiplatform==1.111.0
google-auth==2.40.3
google-api-core==2.25.1

# Environment and utilities
python-dotenv==1.1.1

# Core dependencies
pydantic==2.11.7
numpy==2.3.2
requests==2.32.5